[DEV] Initiate pathfinding array
