+++ editmap2/init_sprite.c
@@ -1252,6 +1252,11 @@ struct gui_sprite *init_gui()
     a = SDL_ConvertSurfaceFormat(a, SDL_GetWindowPixelFormat(window), 0);
     to_return->menu_bas = SDL_CreateTextureFromSurface(renderer, a);
     SDL_FreeSurface(a);
+    a = SDL_LoadBMP("img/gui/selected.bmp");
+    SDL_SetColorKey(a, SDL_SCANCODE_Y, SDL_MapRGB(a->format,255,255,255));
+    a = SDL_ConvertSurfaceFormat(a, SDL_GetWindowPixelFormat(window), 0);
+    to_return->selecteur = SDL_CreateTextureFromSurface(renderer, a);
+    SDL_FreeSurface(a);
     return to_return;
 }
