+++ editmap2/init_sprite.h
@@ -205,6 +205,7 @@ struct gui_sprite
     SDL_Texture *plus;
     SDL_Texture *moins;
     SDL_Texture *menu_bas;
+    SDL_Texture *selecteur;
 };
