+++ jeu/initWorld.h
@@ -1,0 +1,10 @@
+#ifndef INIT_WORLD_H
+#define INIT_WORLD_H
+
+#include "shared_var.h"
+#include "init_sprite.h"
+
+void create_array(char *ground_string);
+
+
+#endif /* INIT_WORLD_H */
