+++ jeu/pathfinding.c
@@ -11,135 +11,20 @@ void print_array(struct path *array)
 	printf ("\n\n--------------------\n\n");
 }
 
-void create_array(char *ground_string)
+void createArray(struct personnages *p)
 {
-	int i = 0;
-	sscanf (ground_string, "%d %d", &max_x, &max_y);
-	while (ground_string[i] != '\n')
-		i++;
-	i++;
-	ground_texture = malloc(sizeof(SDL_Texture*)*max_x*max_y);
-	ground_altitude = malloc(sizeof(int)*max_x*max_y);
-	for (int j = 0; j < max_y*max_x; j++)
+	for(int i = 0; i<max_x*max_y; i++)
 	{
-		if (ground_string[i] == 'e')
-		{
-			if (ground_string[i + 1] == 'a')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->ea1;
-				else if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->ea2;
-				else if (ground_string[i + 2] == '3')
-					ground_texture[j] = img->t->ea3;
-			}
-			i += 4;
-		}
-		else if (ground_string[i] == 't')
-		{
-			if (ground_string[i + 1] == 'e')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->te1;
-				else if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->te2;
-				else if (ground_string[i + 2] == '3')
-					ground_texture[j] = img->t->te3;
-			}
-			i += 4;
-		}
-		else if (ground_string[i] == 'h')
-		{
-			if (ground_string[i + 1] == 'e')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->he1;
-				else if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->he2;
-				else if (ground_string[i + 2] == '3')
-					ground_texture[j] = img->t->he3;
-				else if (ground_string[i + 2] == '4')
-					ground_texture[j] = img->t->he4;
-				else if (ground_string[i + 2] == '5')
-					ground_texture[j] = img->t->he5;
-			}
-			i += 4;
-		}
-		else if (ground_string[i] == 's')
-		{
-			if (ground_string[i + 1] == 'a')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->sa1;
-				else if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->sa2;
-				else if (ground_string[i + 2] == '3')
-					ground_texture[j] = img->t->sa3;
-			}
-			i += 4;
-		}
-		else if (ground_string[i] == 'b')
-		{
-			if (ground_string[i + 1] == 'l')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->bl1;
-				else if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->bl2;
-				else if (ground_string[i + 2] == '3')
-					ground_texture[j] = img->t->bl3;
-			}
-			i += 4;
-		}
-		else if (ground_string[i] == 'n')
-		{
-			if (ground_string[i + 1] == 'e')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->ne1;
-				else if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->ne2;
-				else if (ground_string[i + 2] == '3')
-					ground_texture[j] = img->t->ne3;
-			}
-			i += 4;
-		}
-		else if (ground_string[i] == 'g')
-		{
-			if (ground_string[i + 1] == 'r')
-			{
-				if (ground_string[i + 2] == '1')
-					ground_texture[j] = img->t->gr1;
-				if (ground_string[i + 2] == '2')
-					ground_texture[j] = img->t->gr2;
-			}
-			i += 4;
-		};
-		ground_altitude[j] = atoi(ground_string + i);
-		while (ground_string[i] != ' ' && ground_string[i] != '\n')
-				i++;
-			i++;
+		p->chemin[i].already = 0;
+		p->chemin[i].walkable = (p->eau == 2) || (p->eau == 0 && ground_texture[i] != img->t->ea1 && ground_texture[i] != img->t->ea2 && ground_texture[i] != img->t->ea3) || (p->eau != 0 && ground_texture[i] == img->t->ea1 && ground_texture[i] == img->t->ea2 && ground_texture[i] == img->t->ea3);
 	}
-	i++;
+
+	for(int i = 0; i<max_x*max_y; i++)
+	{
+		
+	}
 }
 
-char *actualise_array(struct linked_list *list)
-{
-	int s = max_x * max_y + 1;
-	char *array_cp = malloc(s);
-	for (int i = 0; i < s; i++)
-		array_cp[i] = 0;
-	for (struct linked_list *ll = list; ll != NULL; ll = ll->next)
-	{
-		if (strcmp(ll->p->skin, "arbre1") == 0)
-		{
-			int xx = ((int)round(ll->p->x) - (int)round(ll->p->x) % 25) / 25;
-			int yy = ((int)round(ll->p->y) - (int)round(ll->p->y) % 25) / 25;
-			array_cp[yy * max_x + xx] = -3;
-		}
-	}
-	return array_cp;
-}
 
 int findpath(struct personnages *p, char *array)
 {
