+++ jeu/pathfinding.h
@@ -8,19 +8,16 @@
 #include "map.h"
 #include "shared_var.h"
 
-void create_array(char *ground);
-
-
 struct path
 {
     int value;
-    char type;
+    char walkable;
     char already;
     int prev;
 };
 
 int can_walk(char type, struct personnages *p);
-char *actualise_array(struct linked_list *list);
+void createArray(struct personnages *p);
 int find_next(struct path *array);
 void generate_around(struct path *array, int src, struct personnages *p);
 int findpath(struct personnages *p, char *array);
